+++ conda/platform_windows.go
@@ -8,6 +8,7 @@ import (
 	"golang.org/x/sys/windows/registry"
 
 	"github.com/robocorp/rcc/common"
+	"github.com/robocorp/rcc/pathlib"
 	"github.com/robocorp/rcc/pretty"
 	"github.com/robocorp/rcc/settings"
 	"github.com/robocorp/rcc/shell"
@@ -46,7 +47,12 @@ func CondaEnvironment() []string {
 }
 
 func BinMicromamba() string {
-	return common.ExpandPath(filepath.Join(common.BinLocation(), "micromamba.exe"))
+	location := common.ExpandPath(filepath.Join(common.MicromambaLocation(), MicromambaVersionNumber))
+	err := pathlib.EnsureDirectoryExists(location)
+	if err != nil {
+		pretty.Warning("Problem creating %q, reason: %v", location, err)
+	}
+	return common.ExpandPath(filepath.Join(location, "micromamba.exe"))
 }
 
 func CondaPaths(prefix string) []string {